Type refactors
