--- service/remote-api/nft-api/models/know-the-characters-response-model.ts
@@ -1,19 +1,0 @@
-interface KnowTheCharactersResponseModel {
-  message: string
-  code: string
-  data: KnowTheCharactersCharacterModel[]
-}
-
-interface KnowTheCharactersCharacterModel {
-  level?: number
-  role: string
-  name: string
-  description: string
-  startIndex: number
-  image: string
-  siblings: number
-  rarity: number
-  type?: string
-}
-
-export { KnowTheCharactersCharacterModel, KnowTheCharactersResponseModel }
