+++ src/views/nft/know-the-characters/NftKnowTheCharacters.astro
@@ -1,5 +1,5 @@
 ---
-import { NftApi } from "../../../../service/remote-api/nft-api/nft-api";
+import { NftApi } from "../../../../service/remote-api/nft-api";
 import Button from "../../../elements/buttons/Button.astro";
 
 const response = await NftApi.knowYourCharacters();
