+++ types/nft-api.d.ts
@@ -1,0 +1,11 @@
+interface NftCharacterModel {
+  level?: number
+  role: string
+  name: string
+  description: string
+  startIndex: number
+  image: string
+  siblings: number
+  rarity: number
+  type?: string
+}
